Added Player Death, Updated Cleaning and Attacking
The player now dies if hit by an enemy attack, at the moment they just lose the ability to do anything, should add death screen and animations in the future. Updated how attacks work, still uses the x key, doesn't have to have an enemy in range. Cleaning moved to the v key. Enemy corpses remove themselves after a bit now, HOWEVER, if the player attacks an enemy while the enemy is dead but before the body disappears, it resets the time until removal and spawns another set of blood particles. Left it in because I thought it might fit into the game.